+++ playlet-web/src/lib/Screens/Settings/SettingControls/EditHomeScreenControl.svelte
@@ -12,7 +12,6 @@
   const textSizes = ["text-2xl", "text-lg", "text-base", "text-sm", "text-xs"];
 
   export let displayText: string = "";
-  // svelte-ignore unused-export-let
   export let key: string = "";
   export let description: string = "";
   export let level: number = 0;
@@ -92,7 +91,7 @@
   <div class={textSizes[level]}>{$translate(displayText)}</div>
   <div class="text-xs text-gray-500">{@html $translate(description)}</div>
   <button class="btn m-1" on:click={openHomeScreenEditor}
-    >{$translate("Edit home screen")}</button
+    >{$translate(displayText)}</button
   >
 </div>